fix eslint errors

+++ client/src/actions/collections.js
@@ -87,52 +87,43 @@ export const updateCollections = (checked, alias, itemId, itemData) => (dispatch
   }
 };
 
-export const getCollectionsFromCategory = (category) => (dispatch) => new Promise((resolve, reject) => {
-  axios.get(FECTH_COLLECTION_BY_CATEGORY_URL(category))
-    .then(({ data }) => {
-      dispatch({
-        type: GET_COLLECTIONS_FROM_CATEGORY,
-        payload: data,
-      });
-
-      resolve(data.length);
-    })
-    .catch((err) => reject(err));
-});
-
-export const createCollection = (userId, data) => (dispatch) => new Promise((resolve, reject) => {
-  axios.post(CREATE_COLLECTION_URL(userId), { collection: data })
-    .then(({ data }) => {
-      if (data.success) {
+export const getCollectionsFromCategory = (category) => (dispatch) => new Promise(
+  (resolve, reject) => {
+    axios.get(FECTH_COLLECTION_BY_CATEGORY_URL(category))
+      .then(({ data }) => {
         dispatch({
-          type: CREATE_COLLECTION,
-          payload: data.item,
+          type: GET_COLLECTIONS_FROM_CATEGORY,
+          payload: data,
         });
-      } else throw new Error(data.message);
-    })
-    .catch((err) => {
-      reject(err);
-    });
-});
 
-export const updateCollection = (alias, values) => (dispatch) => {
-  axios.post(UPDATE_COLLECTION_URL(alias), values)
-    .then(({ data }) => {
-      // if (data.success) {
-      //   dispatch({
-      //     type: UPDATE_COLLECTION,
-      //     payload: data.item
-      //   })
-      // } else throw new Error(data.message)
-    })
-    .catch((err) => {
-      console.log(err);
-    });
-};
+        resolve(data.length);
+      })
+      .catch((err) => reject(err));
+  },
+);
+
+export const createCollection = (userId, collection) => (dispatch) => new Promise(
+  (resolve, reject) => {
+    axios.post(CREATE_COLLECTION_URL(userId), { collection })
+      .then(({ data }) => {
+        if (data.success) {
+          dispatch({
+            type: CREATE_COLLECTION,
+            payload: data.item,
+          });
+        } else throw new Error(data.message);
+      })
+      .catch((err) => {
+        reject(err);
+      });
+  },
+);
 
 export const deleteCollection = (collectionId) => (dispatch) => new Promise((resolve, reject) => {
+  if (!window.confirm('Delete full collection?')) return false;
+
   axios.delete(DELETE_COLLECTION_URL(collectionId))
-    .then((res) => {
+    .then(() => {
       dispatch({
         type: DELETE_COLLECTION,
         payload: collectionId,
@@ -143,16 +134,15 @@ export const deleteCollection = (collectionId) => (dispatch) => new Promise((res
     });
 });
 
-export const getGenres = (collection) => (dispatch) => {
-  axios.get(`${MOVIE_API_BASEURL}/genre/${collection}/list?api_key=${MOVIE_API_KEY}&language=${lang}`)
+export const getGenres = () => (dispatch) => {
+  axios.get(`${MOVIE_API_BASEURL}/genre/movie/list?api_key=${MOVIE_API_KEY}&language=${lang}`)
     .then((res) => {
       const { genres } = res.data;
       axios.get(`${MOVIE_API_BASEURL}/genre/tv/list?api_key=${MOVIE_API_KEY}&language=${lang}`)
-        .then((res) => {
-          // genres.push(res.data.genres)
+        .then(({ data }) => {
           dispatch({
             type: GET_GENRES,
-            genres: [...genres, ...res.data.genres],
+            genres: [...genres, ...data.genres],
           });
         });
     })
